+++ 09-TusLibros.com/Java/src/tusLibros/InterfaceRest.java
@@ -10,7 +10,7 @@ public class InterfaceRest {
 	public static final String CONTRASENIA_INVALIDA = "La contraseña es invalida";
 	public static final String NO_SE_ENCUENTRA_CARRITO = "No se encuentra el carrito buscado";
 	
-	public InterfaceRest(Map<Object, Integer> catalogo) {
+	public InterfaceRest(Map<Object, Double> catalogo) {
 		auten = new AutenticadorCliente(catalogo);
 	}
